+++ example/30-cloudprofile.yaml
@@ -25,7 +25,7 @@ spec:
 #   machineType: # optional -> will use the machine with the lowest amount of CPUs
 #     name: n1-standard-2
 # ---
-# capabilities: # CloudProfileCapabilities feature gate enabled
+# machineCapabilities: # CloudProfileCapabilities feature gate enabled
 # - name: architecture
 #   values: [amd64,arm64]
   kubernetes:
